+++ importmap.js
@@ -1,0 +1,16 @@
+export const importmap = {
+	'imports': {
+		'@shgysk8zer0/kazoo/': 'https://unpkg.com/@shgysk8zer0/kazoo@0.1.1/',
+		'@shgysk8zer0/polyfills': 'https://unpkg.com/@shgysk8zer0/polyfills@0.1.2/all.min.js',
+		'@shgysk8zer0/polyfills/': 'https://unpkg.com/@shgysk8zer0/polyfills@0.1.2/',
+		'@shgysk8zer0/konami': 'https://unpkg.com/@shgysk8zer0/konami@1.0.10/konami.js',
+		'@shgysk8zer0/jswaggersheets': 'https://unpkg.com/@shgysk8zer0/jswaggersheets@1.0.3/swagger.js',
+		'@shgysk8zer0/components/': '../',
+		'@kernvalley/components/': '../krv/',
+		'leaflet': 'https://unpkg.com/leaflet@1.9.3/dist/leaflet-src.esm.js',
+		'marked': 'https://unpkg.com/marked@5.1.0/src/marked.js',
+		'marked-highlight': 'https://unpkg.com/marked-highlight@2.0.1/src/index.js',
+		'highlight.js': 'https://unpkg.com/@highlightjs/cdn-assets@11.8.0/es/highlight.min.js'
+	},
+	scope: {},
+};