+++ test/js/components.js
@@ -15,3 +15,4 @@ import '../../button/share-to.js';
 import '../../krv/events.js';
 import '../../mark-down.js';
 import '../../github/readme.js';
+import '../../toast.js';